+++ sdk/python pip/xr2_sdk/__init__.py
@@ -1,13 +1,22 @@
 from .client import xR2Client, AsyncxR2Client
-from .models import PromptContentResponse, EventResponse, CheckAPIKeyResponse, Response
+from .models import (
+    PromptContentResponse,
+    RenderedPrompt,
+    EventResponse,
+    CheckAPIKeyResponse,
+    Response,
+    VariableError,
+)
 
 __all__ = [
     "xR2Client",
     "AsyncxR2Client",
     "Response",
     "PromptContentResponse",
+    "RenderedPrompt",
     "EventResponse",
     "CheckAPIKeyResponse",
+    "VariableError",
 ]
 
-__version__ = "0.2.5"
+__version__ = "0.3.0"